add access control

+++ routes/articles.js
@@ -1,11 +1,13 @@
 const express = require('express');
 const router = express.Router();
 
-// Bring in Article Model
+// Article Model
 let Article = require('../models/article');
+// User Model
+let User = require('../models/user');
 
 // Add route
-router.get('/add', (req, resp) => {
+router.get('/add', ensureAuth, (req, resp) => {
     resp.render('add_article', {
         title: 'Add Article',
     });
@@ -13,10 +15,9 @@ router.get('/add', (req, resp) => {
 
 
 // Add Submit Post Route
-router.post('/add', (req, resp) => {
+router.post('/add', async (req, resp) => {
 
     req.checkBody('title', 'Title is required').notEmpty();
-    req.checkBody('author', 'Author is required').notEmpty();
     req.checkBody('body', 'Body is required').notEmpty();
 
     // Get Errors
@@ -30,10 +31,11 @@ router.post('/add', (req, resp) => {
     } else {
         let article = new Article();
         article.title = req.body.title;
-        article.author = req.body.author;
+        // @ts-ignore
+        article.author = req.user.id || null;
         article.body = req.body.body;
 
-        article.save()
+        await article.save()
             .then(() => {
                 req.flash('success', 'Article Added');
                 resp.redirect('/');
@@ -45,25 +47,40 @@ router.post('/add', (req, resp) => {
 // Get Single Route
 router.get('/:id', (req, resp) => {
     Article.findById(req.params.id)
-        .then((art) => resp.render('article', {
-            article: art
-        }))
+        .then((art) => {
+            User.findById(art?.author)
+                .then((user) => {
+                    resp.render('article', {
+                        article: art,
+                        author: user?.name
+                    });
+                });
+        })
         .catch((err) => console.log(err));
 });
 
 // Load Edit Article Route
 
-router.get('/edit/:id', (req, resp) => {
+router.get('/edit/:id', ensureAuth, (req, resp, next) => {
     Article.findById(req.params.id)
-        .then((art) => resp.render('edit_article', {
-            title: 'Edit Article',
-            article: art
-        }))
+        .then((art) => {
+            // @ts-ignore
+            if (req.user?.id == art.author) {
+                resp.render('edit_article', {
+                    title: 'Edit Article',
+                    article: art
+                });
+            } else {
+                req.flash('danger', 'Action not authorized.');
+                resp.redirect('/');
+                next();
+            }
+        })
         .catch((err) => console.log(err));
 });
 
 // Update Submit Post Route
-router.post('/edit/:id', (req, resp) => {
+router.post('/edit/:id', ensureAuth, (req, resp) => {
     let article = {};
     article.title = req.body.title;
     article.author = req.body.author;
@@ -80,15 +97,31 @@ router.post('/edit/:id', (req, resp) => {
 });
 
 // Delete Article
-router.delete('/:id', (req, resp) => {
+router.delete('/:id', ensureAuth, async (req, resp) => {
     let query = { _id: req.params.id };
-    Article.deleteOne(query)
-        .then(() => {
-            resp.send('Success');
-        })
-        .catch((err) => console.log(err));
-
+    await Article.findById(req.params.id)
+        .then((err, article) => {
+            // @ts-ignore
+            if (article.author != req.user?._id) {
+                resp.status(500).send();
+            } else {
+                Article.deleteOne(query)
+                    .then(() => {
+                        resp.send('Success');
+                    })
+                    .catch((err) => console.log(err));
+            }
+        });
 });
 
+// Access Control
+function ensureAuth(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next();
+    } else {
+        req.flash('danger', 'Please login');
+        res.redirect('/users/login');
+    }
+}
 
 module.exports = router;
